@@ -11,29 +11,26 @@
     "lint": "yarn eslint --ext .js,.vue ./resources/js"
   },
   "devDependencies": {
-    "@babel/eslint-parser": "^7.11",
-    "@inertiajs/inertia": "^0.8",
-    "@inertiajs/inertia-vue": "^0.5",
+    "@babel/eslint-parser": "^7.15",
+    "@heroicons/vue": "^1.0",
+    "@inertiajs/inertia": "^0.10",
+    "@inertiajs/inertia-vue3": "^0.5",
     "@inertiajs/progress": "^0.2",
-    "apexcharts": "^3.25",
-    "bootstrap": "^4.5",
-    "bootstrap-icons": "^1.2",
+    "@vue/compiler-sfc": "^3.1",
+    "apexcharts": "^3.28",
+    "autoprefixer": "^10.3",
     "eslint": "^7.7",
     "eslint-config-airbnb-base": "^14.2",
     "eslint-plugin-import": "^2.22",
-    "eslint-plugin-vue": "^6.2",
-    "is-svg": "^4.2.2",
-    "jquery": "^3.5",
+    "eslint-plugin-vue": "^7.14",
     "laravel-mix": "^6.0",
-    "popper.js": "^1.16",
-    "postcss": "^8.1",
-    "resolve-url-loader": "^3.1",
-    "sass": "^1.26",
-    "sass-loader": "^8.0",
-    "vue": "^2.6",
-    "vue-apexcharts": "^1.6.0",
-    "vue-loader": "^15.9",
-    "vue-meta": "^2.4",
-    "vue-template-compiler": "^2.6"
+    "lodash": "^4.17",
+    "luxon": "^1.25",
+    "postcss": "^8.3",
+    "resolve-url-loader": "^4.0",
+    "tailwindcss": "^2.2",
+    "vue": "^3.0",
+    "vue-loader": "^16.4",
+    "vue3-apexcharts": "^1.4"
   }
 }
