@@ -1,43 +1,21 @@
-import Vue from 'vue';
-import VueMeta from 'vue-meta';
-import VueApexCharts from 'vue-apexcharts';
-import { App, plugin } from '@inertiajs/inertia-vue';
+import { DateTime } from 'luxon';
+import { createApp, h } from 'vue';
+import { createInertiaApp, InertiaLink, InertiaHead } from '@inertiajs/inertia-vue3';
 import { InertiaProgress } from '@inertiajs/progress';
+import VueApexCharts from 'vue3-apexcharts';
 
-try {
-    window.Popper = require('popper.js').default;
-    window.$ = window.jQuery = require('jquery');
-
-    require('bootstrap');
-} catch (e) {}
-
-InertiaProgress.init();
-
-Vue.config.productionTip = false;
-Vue.use(plugin);
-Vue.use(VueMeta);
-Vue.use(VueApexCharts);
+window._ = require('lodash');
 
 window.Apex = {
-    theme: {
-        mode: 'dark'
-    },
-
     chart: {
-        background: 'transparent',
-        foreColor: '#b4c9de',
         animations: {
             enabled: false
         }
     },
 
     colors: [
-        '#00bcd4'
+        'rgb(6, 182, 212)'
     ],
-
-    grid: {
-        borderColor: '#74818f'
-    },
 
     xaxis: {
         type: 'datetime'
@@ -53,28 +31,36 @@ window.Apex = {
     }
 };
 
-Vue.component('apexchart', VueApexCharts);
+InertiaProgress.init();
 
-Vue.mixin({
-    methods: {
-        icon(name) {
-            return `${this.$page.props.icons}#${name}`;
-        }
+createInertiaApp({
+    // eslint-disable-next-line import/no-dynamic-require
+    resolve: name => require(`./pages/${name}`),
+    setup({
+        el, App, props, plugin
+    }) {
+        props.titleCallback = title => (title
+            ? `${title} - ${props.initialPage.props.title}`
+            : props.initialPage.props.title);
+
+        createApp({ render: () => h(App, props) })
+            .use(plugin)
+            .use(VueApexCharts)
+            .component('InertiaLink', InertiaLink)
+            .component('InertiaHead', InertiaHead)
+            .mixin({
+                methods: {
+                    date(value) {
+                        const date = DateTime.fromISO(value);
+
+                        if (!date.isValid) {
+                            return value;
+                        }
+
+                        return date.toLocaleString(DateTime.DATETIME_MED);
+                    }
+                }
+            })
+            .mount(el);
     }
 });
-
-const el = document.getElementById('app');
-const initialPage = JSON.parse(el.dataset.page);
-
-new Vue({
-    metaInfo: {
-        titleTemplate: title => (title ? `${title} - ${initialPage.props.title}` : initialPage.props.title)
-    },
-
-    render: h => h(App, {
-        props: {
-            initialPage,
-            resolveComponent: name => import(`./pages/${name}.vue`).then(module => module.default)
-        }
-    })
-}).$mount(el);
